pyproject.toml
[tool.poetry]
name = "fastapi-project-tut"
version = "0.1.0"
description = ""
authors = ["liyabonasaki <liyabonasaki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.115.11"}
uvicorn = "^0.34.0"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-dotenv = "^1.1.0"
prometheus-fastapi-instrumentator = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
